Refactor scripting tests
