+++ schemacrawler-scripting/src/test/java/schemacrawler/test/template/CommandlineTemplateCommandTest.java
@@ -1,0 +1,85 @@
+/*
+========================================================================
+SchemaCrawler
+http://www.schemacrawler.com
+Copyright (c) 2000-2019, Sualeh Fatehi <sualeh@hotmail.com>.
+All rights reserved.
+------------------------------------------------------------------------
+
+SchemaCrawler is distributed in the hope that it will be useful, but
+WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+
+SchemaCrawler and the accompanying materials are made available under
+the terms of the Eclipse Public License v1.0, GNU General Public License
+v3 or GNU Lesser General Public License v3.
+
+You may elect to redistribute this code under any of these licenses.
+
+The Eclipse Public License is available at:
+http://www.eclipse.org/legal/epl-v10.html
+
+The GNU General Public License v3 and the GNU Lesser General Public
+License v3 are available at:
+http://www.gnu.org/licenses/
+
+========================================================================
+*/
+
+package schemacrawler.test.template;
+
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static schemacrawler.test.utility.FileHasContent.*;
+import static schemacrawler.test.utility.ScriptTestUtility.commandLineTemplateExecution;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import schemacrawler.test.utility.DatabaseConnectionInfo;
+import schemacrawler.test.utility.TestAssertNoSystemErrOutput;
+import schemacrawler.test.utility.TestAssertNoSystemOutOutput;
+import schemacrawler.test.utility.TestDatabaseConnectionParameterResolver;
+
+@ExtendWith(TestAssertNoSystemErrOutput.class)
+@ExtendWith(TestAssertNoSystemOutOutput.class)
+@ExtendWith(TestDatabaseConnectionParameterResolver.class)
+public class CommandlineTemplateCommandTest
+{
+
+  @Test
+  public void commandlineFreeMarker(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.ftl")),
+               hasSameContentAs(classpathResource("executableForFreeMarker.txt")));
+  }
+
+  @Test
+  public void commandlineMustache(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.mustache")),
+               hasSameContentAs(classpathResource("executableForMustache.txt")));
+  }
+
+  @Test
+  public void commandlineThymeleaf(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.thymeleaf")),
+               hasSameContentAs(classpathResource("executableForThymeleaf.txt")));
+  }
+
+  @Test
+  public void commandlineVelocity(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.vm")),
+               hasSameContentAs(classpathResource("executableForVelocity.txt")));
+  }
+
+}
